+++ apps/traincascade/CMakeLists.txt
@@ -11,6 +11,7 @@ ocv_include_directories("${CMAKE_CURRENT_SOURCE_DIR}" "${OpenCV_SOURCE_DIR}/incl
 ocv_include_modules(${OPENCV_TRAINCASCADE_DEPS})
 
 set(traincascade_files traincascade.cpp
+  common.cpp common.h
   cascadeclassifier.cpp cascadeclassifier.h
   boostlbp.cpp boostlbp.h
   boost.cpp boost.h features.cpp traincascade_features.h
@@ -19,7 +20,7 @@ set(traincascade_files traincascade.cpp
   HOGfeatures.cpp HOGfeatures.h
   mblbpfeatures.cpp mblbpfeatures.h
   imagestorage.cpp imagestorage.h
-  mblbpimgstorage.cpp mblbpimgstorage.h)
+)
 
 set(the_target opencv_traincascade)
 add_executable(${the_target} ${traincascade_files})
